fix(payments): standardize Paymob money typing and widen callback IDs
- Model integer cents money as long (no int overflow) across CashIn DTOs
- Model CashOutTransaction.Amount as decimal instead of double
- Type CashInCreateIntentionRequest.Amount as long (integer cents)
- Widen CashInCallbackTransaction Id/IntegrationId/ProfileId/Owner/SourceId/
  ParentTransaction and nested order Id to long, matching the query-parameter
  callback shape
- Document the cents unit on the request/entry money properties

+++ src/Headless.Payments.Paymob.CashIn/Models/Callback/CashInCallbackQueryParameters.cs
@@ -28,7 +28,7 @@ public sealed class CashInCallbackQueryParameters
     public required bool Pending { get; init; }
 
     [JsonPropertyName("amount_cents")]
-    public required int AmountCents { get; init; }
+    public required long AmountCents { get; init; }
 
     [JsonPropertyName("success")]
     public required bool Success { get; init; }
@@ -85,10 +85,10 @@ public sealed class CashInCallbackQueryParameters
     public long ProfileId { get; init; }
 
     [JsonPropertyName("merchant_commission")]
-    public int MerchantCommission { get; init; }
+    public long MerchantCommission { get; init; }
 
     [JsonPropertyName("accept_fees")]
-    public int AcceptFees { get; init; }
+    public long AcceptFees { get; init; }
 
     [JsonPropertyName("is_void")]
     public bool IsVoid { get; init; }
@@ -97,10 +97,10 @@ public sealed class CashInCallbackQueryParameters
     public bool IsRefund { get; init; }
 
     [JsonPropertyName("refunded_amount_cents")]
-    public int RefundedAmountCents { get; init; }
+    public long RefundedAmountCents { get; init; }
 
     [JsonPropertyName("captured_amount")]
-    public int CapturedAmount { get; init; }
+    public long CapturedAmount { get; init; }
 
     [JsonPropertyName("updated_at")]
     public DateTime UpdatedAt { get; init; }
